view model create

+++ app/src/main/java/com/example/shoppinglist/domain/ShopItem.kt
@@ -4,7 +4,7 @@ package com.example.shoppinglist.domain
 
 data class ShopItem(
     val name: String,
-    val count: Float,
+    val count: Int,
     val enabled: Boolean,
     var id: Int = UNDEFINED_ID
 ) {
